backend/colevents/requirements.txt
addict==2.2.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0
django-cors-headers==2.4.0
django-extra-fields==1.0.0
django-filter==2.0.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
oauthlib==3.0.1
openapi-codec==1.3.2
psycopg2==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
redis==3.0.1
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
uritemplate==3.0.0
urllib3==1.24.1
